pyproject.toml
[tool.poetry]
name = "fastapi_task_manager"
version = "0.1.0"
description = "Task manager on FastApi"
authors = ["Safintim <timurtlt96@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = {extras = ["all"], version = "^0.65.1"}
pydantic = "^1.8.2"
python-dotenv = "^0.17.1"
SQLAlchemy = "^1.4.17"
python-jose = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
wemake-python-styleguide = "^0.15.2"
pytest-cov = "^2.12.0"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
